Move ReadLineEvent, RetryStrategy and ShellRunnerStaticInfo to new files.
