+++ Common/ShellRunner/IShellRunner.cs
@@ -17,20 +17,4 @@ namespace Common
         int RunInDirectory(string path, string commandWithArguments);
         int RunInDirectory(string path, string commandWithArguments, TimeSpan timeout, RetryStrategy retryStrategy = RetryStrategy.IfTimeout);
     }
-    
-    // TODO (DonMorozov): избавиться от класса, потребуется рефакторинг UsagesBuild
-    public static class ShellRunnerStaticInfo
-    {
-        public static string LastOutput { get; set; }
-    }
-    
-    public delegate void ReadLineEvent(string content);
-    
-    // TODO (DonMorozov): реализация выбора стратегии через enum - не самый красивый путь, для более красивого решения потребуются относительно масштабные изменения
-    public enum RetryStrategy
-    {
-        None,
-        IfTimeout,
-        IfTimeoutOrFailed
-    }
 }
