+++ Common/ShellRunner/ReadLineEvent.cs
@@ -1,0 +1,4 @@
+namespace Common
+{
+    public delegate void ReadLineEvent(string content);
+}
