+++ Common/ShellRunner/ShellRunnerStaticInfo.cs
@@ -1,0 +1,8 @@
+namespace Common
+{
+    // TODO (DonMorozov): избавиться от класса, потребуется рефакторинг UsagesBuild
+    public static class ShellRunnerStaticInfo
+    {
+        public static string LastOutput { get; set; }
+    }
+}
